improved database, added constraints

+++ checker/src/main/kotlin/domain/TestCasesRepository.kt
@@ -4,6 +4,6 @@ import com.sushkpavel.domain.model.TestCase
 import com.sushkpavel.domain.model.TestCaseDTO
 
 interface TestCasesRepository {
-    suspend fun getTestCasesByTaskId(taskId: Int): List<TestCase>
+    suspend fun getTestCasesByTaskId(taskId: Long): List<TestCase>
     suspend fun postTestCases(testCaseDTO: TestCaseDTO): Int
 }
