+++ submission/src/main/java/com/sushkpavel/domain/model/TestResult.kt
@@ -1,7 +1,6 @@
 package com.sushkpavel.domain.model
 
 import com.sushkpavel.utils.InstantSerializer
-import kotlinx.serialization.Contextual
 import kotlinx.serialization.Serializable
 import java.time.Instant
 
@@ -10,9 +9,9 @@ data class TestResult(
     val id: Int,
     val userId: Int,
     val submissionId: Int,
-    val taskId: Int,
+    val taskId: Long,
     val actualResult: String,
     val success: Boolean,
     @Serializable(with = InstantSerializer::class)
-    val createdAt : Instant
+    val createdAt: Instant
 )
